fix(web-ui): resolve lint errors and warnings
- Separate shared helpers and contexts from component exports
  to preserve React Fast Refresh boundaries
- Update consumers and tests to use the extracted modules
- Define BtwSessionPanelContent before its wrapper
- Remove an unused ESLint directive and unnecessary export
