+++ src/web-ui/src/app/components/NavPanel/components/DeviceArtwork.tsx
@@ -1,19 +1,7 @@
 import type { DeviceOverviewDevice } from '../deviceInterconnectionOverview';
 import macbookAir from '../assets/macbook-air.png';
 
-type DeviceArtworkKind = 'device' | 'server' | 'macbook-air';
-
-export function getDeviceArtworkKind(
-  device: Pick<DeviceOverviewDevice, 'kind' | 'name'>,
-): DeviceArtworkKind {
-  if (device.kind === 'execution-host') return 'server';
-  // Device names can identify a model, but the controller's OS cannot identify
-  // a remote machine. Unrecognized names deliberately use neutral artwork.
-  if (device.kind === 'desktop' && /\bmacbook[\s._-]*air\b/i.test(device.name)) {
-    return 'macbook-air';
-  }
-  return 'device';
-}
+import { getDeviceArtworkKind } from './deviceArtworkKind';
 
 export function DeviceArtwork({ device }: { device: DeviceOverviewDevice }) {
   const artwork = getDeviceArtworkKind(device);
