+++ src/web-ui/src/app/components/SceneTopBar/index.ts
@@ -3,5 +3,5 @@ export {
   SceneChromeContribution,
   SceneChromeHost,
   SceneChromeProvider,
-  useSceneChromeContext,
 } from './SceneChrome';
+export { useSceneChromeContext } from './sceneChromeContext';
